requirements.txt
Flask>=3.0.0
gunicorn>=20.1.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.13.0
requests>=2.32.0
apscheduler>=3.10.4
python-dateutil>=2.9.0
ijson>=3.2.0
openai>=1.14